Re-implement LGP parsing, reorganize significantly
LGP parsing is now done with a simple `&[u8]` slice instead of
`Read + Seek`. This allows us to use some zero-copy techniques by
keeping references into that `&[u8]` chunk for string-like data.

---

It became clear to me after diving back into the research for this
project that I am going to need to parse a lot more in-depth data than
just LGP and HRC files. To accommodate this, I've restructured modules
in a way that I believe should be sufficient.

In the future, maybe I will expand this module into its own crate, that
handles dealing with all of the FF7-PC file types? Who knows. It could
be a versatile library for those wishing to create modding tools with
Rust.